Port Pentagod over to Castro's framework, implement a bitboard.
This is a work in progress, there's quite a bit left to do.
The bitboard works, as does the MCTS player. There is no negamax nor pns players/solvers yet.
There is no way to change parameters, other than recompile, and many of those are untested.
There is a basic scoring function, but it can be improved with more patterns and tuning of constants.
Now uses GTP instead of tournament.cpp or http.
The move format is different.